app10/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
sqlalchemy==2.0.12
psycopg2-binary==2.9.6
numpy==1.24.3
scikit-learn==1.2.2
joblib==1.2.0
pandas==2.0.1
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.14.1
python-multipart==0.0.6
python-dotenv==1.0.0
celery==5.2.7
redis==4.5.5
flower==1.2.0
httpx==0.24.0
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-cov==4.1.0
black==23.3.0
flake8==6.0.0
isort==5.12.0
mypy==1.3.0
gunicorn==20.1.0
prometheus-client==0.16.0
sentry-sdk==1.25.1
python-json-logger==2.0.7 